Added significant infrastructure to support Cursor
In particular, the `Cursor<T>` struct provides a generic stream over
deserializable `T`. A cursor can be constructed by providing the URL to
`GET` and will automatically pull further pages down on-demand while
respecting page limits.

Note that there is currently no way to distinguish between "not ready
because response not received" and "not ready because we hit the rate
limit and can't move for 14:59". For the case where you eventually want
to consume the entire stream, this is fine. However, this is less than
ideal.